% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGUIstatus.R
\name{GetGUIstatus}
\alias{GetGUIstatus}
\title{Function to get activation status of IsoCorrectoR GUI}
\usage{
GetGUIstatus()
}
\value{
Returns the activation status of the IsoCorrectoR GUI. FALSE represents closed and TRUE represents active state. Returns NULL if GUI has not been started.
}
\description{
The function \code{GetGUIstatus} can be used in scripts to get the activation status of the IsoCorrectoR GUI (logical, FALSE for closed, TRUE for active).
It is required if the GUI is to be started e.g. via a bash script without manually starting an R session.
}
\examples{

# this function is never called directly but BiocCheck() requires a runnable example.

status<-GetGUIstatus()

}
