% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoCorrectionGUI.R
\name{IsoCorrectionGUI}
\alias{IsoCorrectionGUI}
\title{Graphical User Interface for IsoCorrectoR}
\usage{
IsoCorrectionGUI()
}
\value{
Calls internal function to display IsoCorrectoR's Graphical User Interface
}
\description{
Graphical User Interface for IsoCorrectoR
}
\examples{

 IsoCorrectionGUI()

}
